JWT-Marketplace-Kelompok 4/requirements.txt
Flask==2.3.3
Flask-Cors==3.0.10
PyJWT==2.8.0
python-dotenv==1.0.0
